API for app in homepage
Chỉ xong phần read của database, sẽ bổ sung thêm

+++ WG26/src/main/java/tdtu/edu/un/WG26/Repository/AppRepository.java
@@ -1,18 +1,11 @@
 package tdtu.edu.un.WG26.Repository;
 
-import java.util.List;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import tdtu.edu.un.WG26.Model.App;
 
 @Repository
-public interface AppRepository extends JpaRepository<App, Long>{
-
-//	List<App> get();
-//	
-//	App get(Integer id);
-//	
-//	void save(App app);
+public interface AppRepository extends JpaRepository<App, Integer>{
+	
 }
